requirements.txt
flask
flask-cors
face_recognition
requests
pillow
